chore: edit a gist

+++ lib/elixir_gist/gists.ex
@@ -4,6 +4,7 @@ defmodule ElixirGist.Gists do
   """
 
   import Ecto.Query, warn: false
+  alias ElixirGist.Accounts.User
   alias ElixirGist.Repo
 
   alias ElixirGist.Gists.Gist
@@ -68,10 +69,16 @@ defmodule ElixirGist.Gists do
       {:error, %Ecto.Changeset{}}
 
   """
-  def update_gist(%Gist{} = gist, attrs) do
-    gist
-    |> Gist.changeset(attrs)
-    |> Repo.update()
+  def update_gist(%User{} = user, attrs) do
+    gist = Repo.get!(Gist, attrs["id"])
+
+    if(user.id === gist.user_id) do
+      gist
+      |> Gist.changeset(attrs)
+      |> Repo.update()
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
@@ -86,8 +93,15 @@ defmodule ElixirGist.Gists do
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_gist(%Gist{} = gist) do
-    Repo.delete(gist)
+  def delete_gist(%User{} = user, gist_id) do
+    gist = Repo.get!(Gist, gist_id)
+
+    if user.id == gist.user_id do
+      Repo.delete(gist)
+      {:ok, gist}
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
@@ -153,6 +167,8 @@ defmodule ElixirGist.Gists do
     |> Repo.insert()
   end
 
+  @spec update_saved_gist(SavedGist.t(), map()) ::
+          {:ok, SavedGist.t()} | {:error, Ecto.Changeset.t()}
   @doc """
   Updates a saved_gist.
